Optimize material_properties::add(str, vec, double)
The previous version created a new vector matching the length of the one
passed in in the second position, and containing N copies of the double.

The new version creates a vector of length 2, containing only the first
and last entries in the original vector.

As the values are interpolated between the points, and they are all the
same value, there is no point having intermediate points in the lookup
table.