Implement and test `n4::map` (template argument needed)
This implementation requires the following syntax
```
auto input_values = ... any iterable;

n4::map<T>(f, input_values);
```

where you have to specify the output type of `f` in the template
argument of `n4::map` (`T`). We would like to remove the need to
specify this argument, but it's not obvious how this can be done (in c++).