Explicitly delete rm_instance in one test
This concerns the trivial app test. It is likely that it is
fundamentally different from the other tests in some way, but we can't
quite put our finger on it, and we are already in third or fourth order
problem territory, so we don't really have the time or energy to analyse
it any deeper.

Without this delete, the test was giving these warnings:

> WARNING - Attempt to delete the physical volume store while geometry closed !
> WARNING - Attempt to delete the logical volume store while geometry closed !
> WARNING - Attempt to delete the solid store while geometry closed !
> WARNING - Attempt to delete the region store while geometry closed !

The delete makes these warnings go away.

We did try to convert

```c++
static n4::run_manager* n4::run_manager::rm_instance
```

into an `unique_ptr` in order to get automatic deletion, but we must
have messed something up, because it's giving lots of segfaults.