Switch to C++20 standard in nain4 and client project
This is motivated by the use of designated initializers to set the
optional state, temperature and pressure when creating materials.

Even though it's a C++20 feature, it's already present as an extension
in earlier versions of both clang and gcc. But, without the explicit use
of C++20, its use does generate warnings.