Exclude gcc on MacOS from GHA action
On MacOS the `gcc` command uses clang, but in a version that is less
obvious to pin.

So it is both pointless and prone to spurious errors because of use of
an unexpected compiler version. We've already had at least one example
of this already.

Yes, it is possible to install and use a real gcc on MacOS, but nobody
in our environment is going to do it, and if they do then it's their
problem.