Replace `axis` with `rotate`
The problem with axis is that the resulting rotated `x` and `y` axes
are not obvious. By enforcing a sequence of x,y,z rotations, the
result should be more clear. A method rotate that takes a rotation
matrix is also available. It's up to the user to understand their
rotation matrix.