Find G4 data packages via path globbing
At some point these paths started to contain version numbers. Besides
having to update them every time we bump our dependencies, it creates a
hustle for our clients, which may use a different version of nixpkgs.
This solution takes advantage of the fact that there is only one
subdirectory in the package output to fetch it using path globbing.