Use globbing to avoid version numbers in paths
At some point these paths started including version numbers, which is
annoying and makes it difficult to propagate to clients. This solution
takes advantage of the fact that there is only one subdirectory to fetch
it using file globbing.

Even though this script is not used (#200), it will hopefully become
relevant if that is fixed.